+++ lib/pick_or_save.dart
@@ -2,7 +2,7 @@ export 'src/pick_or_save.dart' show PickOrSave;
 
 export 'src/pick_or_save_method_channel.dart'
     show
-        FilePickingType,
+        PickerType,
         FilePickerParams,
         FileSavingType,
         SaveFileInfo,